Test the way in the app actually takes
The first version of this test built its own route tree and mounted the
page on it, which is the one arrangement that cannot fail the way the
app does: registration and the code-split import are both stepped over,
so a page nobody can reach still reads as reachable.

Each link is checked against the real article instead. The generated
tree has to resolve /communities to the directory's route; the href the
rail renders has to be an address that same tree resolves there; and the
page comes from the route's own component, loaded through the dynamic
import it is declared with, so a moved page or a renamed export fails
here rather than at a click.
